# start implementation in typescript