--- a/.eslintrc
+++ b/.eslintrc
@@ -1,3 +1,0 @@
-{
-  extends: "eslint-config-lob"
-}