--- a/bench/index.ts
+++ b/bench/index.ts
@@ -21,8 +21,9 @@ import * as b from 'benny'
 import * as fs from 'fs'
 import * as path from 'path'
 
-const calipers = require('../lib')('png')
+import * as calipers from '../src'
 
+const COUNT = 5
 const ONE_BILLION = 1000000000
 
 const fixtures = path.resolve(__dirname, '../test/fixtures')
@@ -32,23 +33,27 @@ let files = fs.readdirSync(fixtures)
 files = files.filter(function (f) { return f.endsWith('.png') })
 
 // Create a benchmark for each image.
-const benchmarks = files.map(function (f) {
+const benchmarks: any[] = []
+files.forEach(function (f) {
   const [width, height, type] = f.split(/x|\./)
   const filePath = path.resolve(fixtures, f)
-  return b.add(`${type}-${width}x${height}`, async () => {
-    await calipers.measure(filePath)
+  const bench = b.add(`${type}-${width}x${height}`, async () => {
+    await calipers.measure(filePath, calipers.Format.PNG)
   })
+  for (let i = 0; i < COUNT; i++) {
+    benchmarks.push(bench)
+  }
 })
 
 // Run the benchmark suite.
 b.suite(
-  'calipers',
+  'calipers ts',
   ...benchmarks,
   b.cycle((result) => {
     // Print the results in a format that is compatible with benchstat.
     const name = result.name
     const count = result.samples
     const nsPerOp = result.details.mean * ONE_BILLION
-    return `BenchmarkCalipers/${name}\t${count}\t${nsPerOp} ns/op`
+    console.log(`BenchmarkCalipers/${name}\t${count}\t${nsPerOp} ns/op`)
   })
 )