--- a/jest.config.json
+++ b/jest.config.json
@@ -1,0 +1,9 @@
+{
+  "globals": {
+    "ts-jest": {
+      "tsconfig": "tsconfig.test.json"
+    }
+  },
+  "preset": "ts-jest",
+  "testEnvironment": "node"
+}