--- a/package.json
+++ b/package.json
@@ -4,7 +4,8 @@
   "description": "The fastest Node.js library for measuring image and PDF dimensions.",
   "main": "lib/index.js",
   "scripts": {
-    "bench": "ts-node bench",
+    "bench": "node bench/index.js",
+    "bench:ts": "ts-node bench/index.ts",
     "b": "tsc bench/test",
     "build": "tsc",
     "test": "jest",
@@ -44,6 +45,7 @@
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
     "jest": "26.x.x",
+    "ts-jest": "^26.4.4",
     "ts-node": "^9.0.0",
     "typescript": "^4.0.5"
   },