--- a/test/detect.test.js
+++ b/test/detect.test.js
@@ -2,7 +2,6 @@
 
 var fs      = require('fs');
 var path    = require('path');
-var expect  = require('chai').expect;
 var Promise = require('bluebird');
 var popen   = Promise.promisify(fs.open);
 var detect  = require('../lib/detect');
@@ -29,14 +28,14 @@ describe('detect', function () {
       return detect(fd, [fakeFalsePlugin, fakeTruePlugin]);
     })
     .then(function (plugin) {
-      expect(plugin).to.eql(fakeTruePlugin);
+      expect(plugin).toBe(fakeTruePlugin);
     });
   });
 
   it('should throw an error for an unsupported file type', function () {
     return popen(txtPath, 'r')
     .then(function (fd) {
-      return expect(detect(fd, [fakeFalsePlugin])).to.be.rejectedWith(Error);
+      return expect(detect(fd, [fakeFalsePlugin])).rejects.toThrow(Error);
     });
   });