--- a/test/setup.test.js
+++ b/test/setup.test.js
@@ -1,5 +1,0 @@
-'use strict';
-
-var chai = require('chai');
-
-chai.use(require('chai-as-promised'));